rebased PR from main to retain hugos commits
